Metrics calculation system and refactorisation
Metrics calculus and metrics themselves are implemented into the package metrics. Metrics are calculated (for now) every 10 seconds over a timeframe of 2 minutes for each website. Metrics are flexible as they are types which implement an interface. Refactoring maps of channels into a single channel for the web watcher as we can discriminate MetaResponse with the name of the website which has been added. We could effectively consider each MetaResponse independantly, so the overall increase of memory complexity seems worth it.