Functionnal logging of computed metrics and refactor
There is now a Reporter type which format and logs metrics. The datagrinder file now contains methods to manipulate the shared data but does not know about reporting. The main package is responsible for feeding the data from channels, launch a goroutine with a ticker to compute metrics from data, and report them.