Hooks mecanism and test for alerting logic
Hooks are closures which are meant to be called each time new metrics are calculated. They provide extra logging or working with no side effects on metrics. They do not impact classic reporting and do not have any access to Logger and Writer used by the main application. The alerting logic is implemented as a Hook : it records unavailabilities and recoveries but do not report them (just log them in console). The closure allows to be statefull without changing the main package. The test package contains a test for the alerting logic, which builds a webserver and a configuration from scratch, alters the webserver and confronts the alerting hook (which must detect unavailability and recovery.