Timeout, formatter, logger, metrics and hooks in config
Enhanced configuration file which allows users to customize MicroMon's behaviour. The timeout is a mandatory configuration paramater. Formatter and logger parameters are not required, but when present, changes the behaviour. For the sake of example, an XML formatter and a file logger have been implemented. Metrics and hooks to perform are configured with a friendly name which is linked to concrete instances in metric and hook package.