Enhanced formatter, test modified
Formatters now have a Prefix() and a Suffix() methods, which are meant to be called before reporting groups of metrics for differents websites at the same time. In XML this corresponds to an encapsulating metrics with a special tag. Alerting test has a lower alerting threshold to lower the time to recover from unavailability.